split expression logic
